ECO: Move xls/eco to xls/contrib/eco
Relocate the ECO package from xls/eco to xls/contrib/eco and update references.

This includes Bazel labels, Starlark loads, C++ include paths, Python imports, runfiles paths, and tooling path filters.
