+++ xls/contrib/eco/graph.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "xls/eco/graph.h"
+#include "xls/contrib/eco/graph.h"
 
 #include "absl/container/flat_hash_set.h"
