+++ xls/contrib/eco/gxl_parser.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "xls/eco/gxl_parser.h"
+#include "xls/contrib/eco/gxl_parser.h"
 
 #include <algorithm>
 #include <cerrno>
