+++ xls/contrib/eco/ir_patch_gen.cc
@@ -29,9 +29,9 @@
 #include "absl/strings/str_split.h"
 #include "absl/strings/string_view.h"
 #include "absl/strings/strip.h"
-#include "xls/eco/ged.h"
-#include "xls/eco/graph.h"
-#include "xls/eco/ir_patch.pb.h"
+#include "xls/contrib/eco/ged.h"
+#include "xls/contrib/eco/graph.h"
+#include "xls/contrib/eco/ir_patch.pb.h"
 #include "xls/ir/ir_parser.h"
 #include "xls/ir/package.h"
 #include "xls/ir/type.h"
