+++ xls/contrib/eco/lap_solver.h
