+++ xls/contrib/eco/patch_ir.h
@@ -28,7 +28,7 @@
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "absl/types/span.h"
-#include "xls/eco/ir_patch.pb.h"
+#include "xls/contrib/eco/ir_patch.pb.h"
 #include "xls/ir/function_base.h"
 #include "xls/ir/node.h"
 #include "xls/ir/op.h"
