+++ xls/contrib/eco/patch_ir_main.cc
@@ -26,8 +26,8 @@
 #include "xls/common/file/filesystem.h"
 #include "xls/common/init_xls.h"
 #include "xls/common/status/status_macros.h"
-#include "xls/eco/ir_patch.pb.h"
-#include "xls/eco/patch_ir.h"
+#include "xls/contrib/eco/ir_patch.pb.h"
+#include "xls/contrib/eco/patch_ir.h"
 #include "xls/ir/function_base.h"
 #include "xls/ir/ir_parser.h"
 #include "xls/ir/node.h"
