+++ xls/contrib/eco/test/vector_core.rev.x
