+++ xls/contrib/eco/xls_ir_to_networkx.py
@@ -12,31 +12,30 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Test utility that converts XLS IR to a networkx graph."""
+"""Read XLS IR into a networkx graph.
 
+Uses the xls_ir_to_cytoscape utility to read the IR into a JSON format and then
+parses that into a networkx graph.
+"""
+
+import json
 import pathlib
+import subprocess
 
+from absl import logging
 import networkx as nx
 
-from absl.testing import absltest
 from xls.common import runfiles
-from xls.eco import xls_ir_to_networkx
 
-_RISCV_SIMPLE_OPT_IR_PATH = runfiles.get_path(
-    "xls/examples/riscv_simple.opt.ir"
+
+XLS_IR_TO_CYTOSCAPE_PATH = runfiles.get_path(
+    "xls/contrib/eco/xls_ir_to_cytoscape"
 )
 
 
-class XlsIrToNetworkxTest(absltest.TestCase):
-
-  def test_riscv_simple(self):
-    graph = xls_ir_to_networkx.read_xls_ir_to_networkx(
-        pathlib.Path(_RISCV_SIMPLE_OPT_IR_PATH)
-    )
-    self.assertIsInstance(graph, nx.DiGraph)
-    self.assertNotEmpty(graph.nodes())
-    self.assertNotEmpty(graph.edges())
-
-
-if __name__ == "__main__":
-  absltest.main()
+def read_xls_ir_to_networkx(ir_path: pathlib.Path) -> nx.MultiDiGraph:
+  """Reads XLS IR into a networkx graph."""
+  json_text = subprocess.check_output([XLS_IR_TO_CYTOSCAPE_PATH, ir_path])
+  ir_json = json.loads(json_text)
+  logging.vlog(3, "Loaded JSON: %s", ir_json)
+  return nx.cytoscape_graph(ir_json)
